@@ -1,18 +1,13 @@
-import ReactDOM from 'react-dom';
+import React from "react";
+import ReactDOM from "react-dom";
 import {BrowserRouter} from "react-router-dom";
-
 import './assets/css/index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+
+import App from "./App";
 
 ReactDOM.render(
   <BrowserRouter>
     <App/>
   </BrowserRouter>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
